refactor: apply 요청 값 및 헤더 수정
