+++ src/components/common/Header/style.tsx
@@ -37,6 +37,7 @@ export const HeaderIconsArrange = styled.div`
   display: flex;
   justify-content: space-between;
 `;
+
 export const AlertModalTitle = styled.div`
   text-align: center;
 
@@ -55,6 +56,7 @@ export const ModalBody = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
+  align-items: center;
 
   height: 90%;
   padding-top: 30px;
