+++ src/components/common/PageTamplate/index.tsx
@@ -4,14 +4,16 @@ import * as Styled from './style';
 
 import BottomNavigation from '@/components/common/BottomNavigation';
 import Header from '@/components/common/Header';
+import { useFetchMe } from '@/query/user';
 
 const PageTemplate = (props: PropsWithChildren) => {
   const { children } = props;
+  const { me } = useFetchMe();
 
   return (
     <Styled.Root>
       <Styled.Container>
-        <Header />
+        {me && <Header />}
         <Styled.Main>{children}</Styled.Main>
         <BottomNavigation />
       </Styled.Container>
