+++ src/types/locker.ts
@@ -1,3 +1,5 @@
+import { MajorPriorityAnswerRequest } from './major';
+
 export interface LockerRequest {
   major: number;
   building_id: number;
@@ -17,13 +19,13 @@ export interface LockerResponse {
   start_date: string | null;
 }
 
-export interface RequestApplyLocker extends LockerRequest {
+export interface RequestApplyLocker extends Partial<MajorPriorityAnswerRequest> {
   building_id: number;
   major: number;
   user: number;
-  priority_1?: number | boolean | string | null | undefined;
-  priority_2?: number | boolean | string | null | undefined;
-  priority_3?: number | boolean | string | null | undefined;
+  // priority_1_answer?: number | boolean | string | null | undefined;
+  // priority_2_answer?: number | boolean | string | null | undefined;
+  // priority_3_answer?: number | boolean | string | null | undefined;
 }
 
 export type ConvertToShareRequest = Pick<LockerResponse, 'id'> & {
